refactor: 상단물품결제 리팩토링
