+++ GoodBuyUS/src/main/java/dev_final_team10/GoodBuyUS/controller/api/MainPaymentController.java
@@ -33,53 +33,6 @@ public class MainPaymentController {
         return ResponseEntity.ok(responseDto);
     }
 
-    /*@RequestMapping(value = "/success", method = {RequestMethod.GET, RequestMethod.POST})
-    public ResponseEntity<?> handlePaymentSuccess(
-            @RequestParam String paymentKey,
-            @RequestParam UUID orderId,
-            @RequestParam int amount) {
-        try {
-            MainPaymentResponseDto responseDto = mainPaymentService.handlePaymentSuccess(paymentKey, orderId, amount);
-            Long productId = mainPaymentService.getProductIdFromOrder(orderId);
-
-            String redirectUrl = "http://15.164.5.135:8080/products/payment-success/" + productId;
-
-            return ResponseEntity.ok(Map.of(
-                    "productName", responseDto.getProductName(),
-                    "quantity", responseDto.getQuantity(),
-                    "price", responseDto.getPrice(),
-                    "totalAmount", responseDto.getTotalPrice(),
-                    "status", responseDto.getStatus(),
-                    "redirectUrl", redirectUrl
-
-            ));
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(Map.of("error", "결제 성공 처리 중 오류 발생", "message", e.getMessage()));
-        }
-    }*/
-
-    /*@PostMapping("/fail")
-    public ResponseEntity<?> handlePaymentFail(
-            @RequestParam String orderId,
-            @RequestParam String message) {
-        try {
-            // Order ID를 기반으로 Product ID 가져오기
-            Long productId = mainPaymentService.getProductIdFromOrder(UUID.fromString(orderId));
-
-
-            String redirectUrl = "http://15.164.5.135:8080/products/payment-fail/" + productId;
-
-            return ResponseEntity.ok(Map.of(
-                    "status", "fail",
-                    "redirectUrl", redirectUrl
-            ));
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(Map.of("error", "결제 실패 처리 중 오류 발생", "message", e.getMessage()));
-        }
-    }*/
-
     @PostMapping("/approve")
     public ResponseEntity<MainPaymentResponseDto> approvePayment(@RequestBody Map<String, Object> requestBody) {
         String paymentKey = (String) requestBody.get("paymentKey");
